# contracts/04-interoperability/openapi-outline.yaml
openapi: 3.1.0
info:
  title: AI Signal Studio API
  version: 0.1.0
paths:
...
  /api/agent-conversations/{conversation_id}:
    parameters:
      - in: path
        name: conversation_id
        required: true
        schema: {type: string}
    get:
      operationId: getAgentConversation
      summary: Read one conversation and its persisted messages
      responses:
        '200':
          description: Conversation detail
    patch:
      operationId: updateAgentConversation
      summary: Rename or pin a conversation
      responses:
        '200':
          description: Updated conversation
    delete:
      operationId: deleteAgentConversation
      summary: Softly delete a conversation
      responses:
        '200':
          description: Softly deleted conversation
  /api/agent-conversations/{conversation_id}/archive:
    parameters:
      - in: path
        name: conversation_id
        required: true
        schema: {type: string}
    post:
      operationId: archiveAgentConversation
      responses:
        '200':
          description: Archived conversation
  /api/agent-conversations/{conversation_id}/restore:
    parameters:
      - in: path
        name: conversation_id
        required: true
        schema: {type: string}
    post:
      operationId: restoreAgentConversation
      responses:
        '200':
          description: Restored conversation with messages preserved
  /api/agent-conversations/{conversation_id}/turns:
    parameters:
      - in: path
        name: conversation_id
        required: true
        schema: {type: string}
    post:
      operationId: createAgentTurn
      summary: Persist an idempotent Agent turn and start the 0.7.0 graph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message, client_message_id]
              properties:
                message: {type: string, minLength: 1, maxLength: 4000}
                client_message_id: {type: string, minLength: 1, maxLength: 100}
                model_id: {type: [string, 'null']}
                artifact_ids:
                  type: array
                  maxItems: 20
                  items: {type: string}
      responses:
        '202':
          description: Queued turn with its persisted user message
  /api/agent-turns/{turn_id}:
    parameters:
      - in: path
        name: turn_id
        required: true
        schema: {type: string}
    get:
      operationId: getAgentTurn
      summary: Read persisted plan, outcome, duration, blocks, and errors
      responses:
        '200':
          description: Durable Agent turn with Goal, Plan, requested/effective model IDs, an always-present localized result_summary, deduplicated ResultBlocks, evidence-window disclosure, and actionable errors
  /api/agent-turns/{turn_id}/events:
    parameters:
      - in: path
        name: turn_id
        required: true
        schema: {type: string}
      - in: header
        name: Last-Event-ID
        required: false
        schema: {type: integer, minimum: 0}
    get:
      operationId: streamAgentTurnEvents
      summary: Stream persisted ordered events including shared model research, per-step outcomes, and complete validated ResultBlocks, resuming after an event ID
      responses:
        '200':
          description: text/event-stream ending after the durable terminal event
...